Add abstractions for HTTP methods
