+++ .prettierrc.json
@@ -1,3 +1,6 @@
 {
-    "singleQuote": false
+    "singleQuote": false,
+    "bracketSpacing": true,
+    "trailingComma": "es5",
+    "arrowParens": "always"
 }
