+++ package.json
@@ -6,6 +6,7 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "dependencies": {
+    "lodash": "^4.17.21",
     "node-fetch": "^2.6.6",
     "parcel": "^2.7.0",
     "sanctuary": "^3.1.0",
@@ -13,7 +14,7 @@
   },
   "targets": {
     "main": {
-      "optimize":true
+      "optimize": true
     }
   },
   "scripts": {
@@ -25,6 +26,7 @@
     "@jest/globals": "^29.0.3",
     "@parcel/packager-ts": "2.7.0",
     "@parcel/transformer-typescript-types": "2.7.0",
+    "@types/lodash": "^4.14.186",
     "@types/node": "^18.7.18",
     "@types/node-fetch": "^2.6.2",
     "@types/sanctuary": "^3.1.0",
