+++ src/types/fetch.ts
@@ -1,0 +1,1 @@
+export type Primitives = string | number | boolean;
